+++ embassy-net/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+- update embedded-nal-async
+
 ## 0.4 - 2024-01-11
 
 - Update to `embassy-time` v0.3.